+++ NEWS.md
@@ -6,6 +6,8 @@
 
 ## Improvements
 
+* `hdr()` now produces exact results for symmetric distributions (#123).
+* `hdr()` uses more quantiles by default (512 -> 4096) for more accurate HDRs (#123).
 * Transformed distributions now work better for monotonic increasing functions 
   when the domain of the untransformed distribution is R (#129).
 * Added cdf() and quantile() methods for dist_categorical() distributions with